+++ Flutter/accessibility_stats_logger/lib/accessibility_stats_logger.dart
@@ -1,11 +1,7 @@
 import 'accessibility_stats_logger_platform_interface.dart';
 
 class AccessibilityStatsLogger {
-  Future<String?> getPlatformVersion() {
-    return AccessibilityStatsLoggerPlatform.instance.getPlatformVersion();
-  }
-
-  Future<String?> getAccessibilityStats() {
+  Future<Map<String, Object?>> getAccessibilityStats() {
     return AccessibilityStatsLoggerPlatform.instance.getAccessibilityStats();
   }
 }