+++ Flutter/accessibility_stats_logger/lib/accessibility_stats_logger_method_channel.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+import 'dart:convert';
 
 import 'accessibility_stats_logger_platform_interface.dart';
 
@@ -9,16 +10,15 @@ class MethodChannelAccessibilityStatsLogger
   final methodChannel = const MethodChannel('accessibility_stats_logger');
 
   @override
-  Future<String?> getPlatformVersion() async {
-    final version =
-        await methodChannel.invokeMethod<String>('getPlatformVersion');
-    return version;
-  }
-
-  @override
-  Future<String?> getAccessibilityStats() async {
-    final version =
-        await methodChannel.invokeMethod<String>('getAccessibilityStats');
-    return version;
+  Future<Map<String, Object?>> getAccessibilityStats() async {
+    try {
+      final accessibilityString =
+          await methodChannel.invokeMethod<String>('getAccessibilityStats');
+      final accessiblityData =
+          jsonDecode(accessibilityString.toString()) as Map<String, Object?>;
+      return accessiblityData;
+    } catch (e) {
+      return {};
+    }
   }
 }