+++ Flutter/accessibility_stats_logger/lib/accessibility_stats_logger_platform_interface.dart
@@ -17,12 +17,5 @@ abstract class AccessibilityStatsLoggerPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Future<String?> getPlatformVersion() {
-    throw UnimplementedError('platformVersion() has not been implemented.');
-  }
-
-  Future<String?> getAccessibilityStats() {
-    throw UnimplementedError(
-        'getAccessibilityStats() has not been implemented.');
-  }
+  Future<Map<String, Object?>> getAccessibilityStats();
 }