+++ Flutter/accessibility_stats_logger/test/accessibility_stats_logger_method_channel_test.dart
@@ -1,15 +1,13 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:accessibility_stats_logger/accessibility_stats_logger_method_channel.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
-
-  MethodChannelAccessibilityStatsLogger platform = MethodChannelAccessibilityStatsLogger();
   const MethodChannel channel = MethodChannel('accessibility_stats_logger');
 
   setUp(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(
       channel,
       (MethodCall methodCall) async {
         return '42';
@@ -18,10 +16,7 @@ void main() {
   });
 
   tearDown(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
-  });
-
-  test('getPlatformVersion', () async {
-    expect(await platform.getPlatformVersion(), '42');
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, null);
   });
 }