+++ Flutter/accessibility_stats_logger/CHANGELOG.md
@@ -3,3 +3,9 @@
 This Flutter plugin is able to gather Accessibility data from iOS and Android. It gathers data from the ScreenReader, VoiceOver, font size and many more. 
 
 The function ```getAccessibilityStats()``` return a ```String```. You can easily cast to ```Map<String, String>```.
+
+## 1.1.0
+
+This Flutter plugin is able to gather Accessibility data from iOS and Android. It gathers data from the ScreenReader, VoiceOver, font size and many more. 
+
+The function ```getAccessibilityStats()``` now returns a ```Map<String, Object?``` and has error handling. It is coveres its code in unit tests.